==> Scripts/2_1_Hu_homo_coronaries_scPrep.Rmd <==
---
title: "Hu_homo_coronaries_scPrep"
output: html_document
date: "2024-04-06"
editor_options: 
  chunk_output_type: console
---

# Load libraries
```{r}
########### Source: https://zenodo.org/records/6032099

setwd("Human/")

library(Seurat)
library(tidyverse)
library(data.table)
library(scDblFinder)
library(celda)
library(sctransform)
library(cluster)
library(readr)
source("../scRNA_processing_utils.R") # with the modifications from https://github.com/MillerLab-CPHG/Human_athero_scRNA_meta/blob/main/utils/scRNA_processing_utils.R

# Set seed for doublet removal reproducibility
set.seed(1)

```


# Patient2 Coronary Artery 2
## Prepare the count matrix
```{r}
coronary2_p2 = Seurat::Read10X("Hu_coronaries_aortas_pulmonary/cardiac_arteries_processed_data/patient2_CA2/")

# Raw, unfiltered matrix is 33538 genes x 14973 cells
dim(coronary2_p2)

```

## Create Seurat obj
```{r}
# Pre-processing (Create Seurat obj containing doublets)
# Dims of seurat object: 18538 genes x 14916 cells
coronary2_p2_seurat_sct = CreateSeuratObject(counts = coronary2_p2, 
                                      project = "hu_p2_coronary2", 
                                      min.cells = 10, 
                                      min.features = 200)

# before/after removing doublets
# SCT normalize
coronary2_p2_seurat_sct = Seurat_SCT_process(coronary2_p2_seurat_sct, seurat_filter = TRUE,
                                      sample_id = "hu_coronary2_p2", 
                                      study_name = "hu_et_al")
###########
### Finding Best Resolution for integrated object

resolutions_sil_df = calc_sil_scores_for_res(coronary2_p2_seurat_sct, clustering_res = c(0.5, 0.6, 0.7, 0.9, 0.75 ,0.8, 1.0,1.1, 1.4))

max_avg_sil_score(resolutions_sil_df) # Max is 0,5
##########

# Finding the right resolution is quite important for doublet detection and removal
coronary2_p2_seurat_sct =  FindClusters(coronary2_p2_seurat_sct, resolution = 0.8)   

# Visualize clusters
p3_before_QC = DimPlot(coronary2_p2_seurat_sct, reduction = "umap", label = TRUE) +
  ggtitle("Before removing doublets") + custom_theme + level2_annotations_scale

ggsave(filename = "Hu_coronaries_aortas_pulmonary/Hu_coronary2_p2_before_QC.png", width = 7, height = 8)

```

## Detect doublets & ambient RNA
```{r}
# Detect doublets with scDblFinder
# Called 728 doublets
coronary2_p2_doublet_ids = scDblFinder_clusters(coronary2_p2_seurat_sct, nrep = 3)
length(coronary2_p2_doublet_ids)

# First filtering step: Use the vector of consensus doublet IDs output below
coronary2_p2_singlets = setdiff(colnames(coronary2_p2), coronary2_p2_doublet_ids)
length(coronary2_p2_singlets)

coronary2_p2_seurat_sct = CreateSeuratObject(counts = coronary2_p2[, coronary2_p2_singlets], 
                                      project = "hu_p2_coronary2", 
                                      min.cells = 10, 
                                      min.features = 200)

# Clean counts matrix (without doublets) from ambient RNA using the decontX wrapper
coronary2_p2_seurat_sct =  decontX_remove(coronary2_p2_seurat_sct)
head(coronary2_p2_seurat_sct@assays$RNA@counts)
```

```{r}
#after removing doublets
# SCT normalize
coronary2_p2_seurat_sct = Seurat_SCT_process(coronary2_p2_seurat_sct, seurat_filter = TRUE,
                                      sample_id = "hu_coronary2_p2", 
                                      study_name = "hu_et_al")

# Finding the right resolution is quite important for doublet detection and removal
coronary2_p2_seurat_sct =  FindClusters(coronary2_p2_seurat_sct, resolution = 0.8) # max is 0.6

p3_after_QC = DimPlot(coronary2_p2_seurat_sct, reduction = "umap", label = TRUE) +
  ggtitle("After removing doublets + decontX") + custom_theme + level2_annotations_scale
ggsave(filename = "Hu_coronaries_aortas_pulmonary/Hu_coronary2_p2_after_QC.png", width = 7, height = 8)

FeaturePlot(coronary2_p2_seurat_sct, features = c("MYH11", "CNN1", "TNFRSF11B", "ACTA2", 
                                           "CRTAC1", "LUM"), order = TRUE,  pt.size = 0.1)

FeaturePlot(coronary2_p2_seurat_sct, features = c("CD68", "CD14", "CD163", "APOE", 
                                           "S100A8", "IL1B"), order = TRUE,  pt.size = 0.1)

saveRDS(coronary2_p2_seurat_sct, "Hu_coronaries_aortas_pulmonary/hu_coronary2_p2_processed_seurat_obj.rds")

```


# Patient3 Coronary Artery 2
## Prepare the count matrix
```{r}
coronary2_p3 = Seurat::Read10X("Hu_coronaries_aortas_pulmonary/cardiac_arteries_processed_data/patient3_CA2/")

# Raw, unfiltered matrix is 33538 genes x 8091 cells
dim(coronary2_p3)
```
## Create Seurat obj
```{r}
#Pre-processing (Create Seurat obj containing doublets)
# Dims of seurat object: 17861 genes x 7994 cells
coronary2_p3_seurat_sct = CreateSeuratObject(counts = coronary2_p3, 
                                      project = "hu_p3_coronary2", 
                                      min.cells = 10, 
                                      min.features = 200)

# before/after removing doublets
# SCT normalize
coronary2_p3_seurat_sct = Seurat_SCT_process(coronary2_p3_seurat_sct, seurat_filter = TRUE,
                                      sample_id = "hu_coronary2_p3", 
                                      study_name = "hu_et_al")

###########
### Finding Best Resolution for integrated object

resolutions_sil_df = calc_sil_scores_for_res(coronary2_p3_seurat_sct, clustering_res = c(0.5, 0.6, 0.7, 0.9, 0.75 ,0.8, 1.0,1.1, 1.4))

max_avg_sil_score(resolutions_sil_df) # Max is 0.5
##########

# Finding the right resolution is quite important for doublet detection and removal
coronary2_p3_seurat_sct =  FindClusters(coronary2_p3_seurat_sct, resolution = 0.8)   

# Visualize clusters
p5_before_QC = DimPlot(coronary2_p3_seurat_sct, reduction = "umap", label = TRUE) +
  ggtitle("Before removing doublets")
ggsave(filename = "Hu_coronaries_aortas_pulmonary/Hu_coronary2_p3_before_QC.png", width = 7, height = 8)

```
## Detect doublets & ambient RNA
```{r}
# Detect doublets with scDblFinder
# Called 118 doublets
coronary2_p3_doublet_ids = scDblFinder_clusters(coronary2_p3_seurat_sct, nrep = 3)
length(coronary2_p3_doublet_ids)

# First filtering step: Use the vector of consensus doublet IDs output below
coronary2_p3_singlets = setdiff(colnames(coronary2_p3), coronary2_p3_doublet_ids)
length(coronary2_p3_singlets)

# New dims after removing doublets: 17716 genes x 7629 cells
coronary2_p3_seurat_sct = CreateSeuratObject(counts = coronary2_p3[, coronary2_p3_singlets], 
                                      project = "hu_p3_coronary2", 
                                      min.cells = 10, 
                                      min.features = 200)

# Clean counts matrix (without doublets) from ambient RNA using the decontX wrapper
coronary2_p3_seurat_sct =  decontX_remove(coronary2_p3_seurat_sct)
head(coronary2_p3_seurat_sct@assays$RNA@counts)
```

```{r}
# after removing doublets
# SCT normalize
coronary2_p3_seurat_sct = Seurat_SCT_process(coronary2_p3_seurat_sct, seurat_filter = TRUE,
                                      sample_id = "hu_coronary2_p3", 
                                      study_name = "hu_et_al")

# Finding the right resolution is quite important for doublet detection and removal
coronary2_p3_seurat_sct =  FindClusters(coronary2_p3_seurat_sct, resolution = 0.8) #Max is 0.5

p5_after_QC = DimPlot(coronary2_p3_seurat_sct, reduction = "umap", label = TRUE) +
  ggtitle("After removing doublets + decontX")
ggsave(filename = "Hu_coronaries_aortas_pulmonary/Hu_coronary2_p3_after_QC.png", width = 7, height = 8)


FeaturePlot(coronary2_p3_seurat_sct, features = c("MYH11", "CNN1", "TNFRSF11B", "ACTA2", 
                                           "CRTAC1", "LUM"), order = TRUE,  pt.size = 0.1)

FeaturePlot(coronary2_p3_seurat_sct, features = c("CD68", "CD14", "CD163", "TREM2", 
                                           "S100A8", "IL1B"), order = TRUE,  pt.size = 0.1)

saveRDS(coronary2_p3_seurat_sct, "Hu_coronaries_aortas_pulmonary/hu_coronary2_p3_processed_seurat_obj.rds")

```
